File rearrangement and html to jsx conversion update
A rearrangement of the filesystem was done for a clear overview. In that process new files were added and changed based on existing files such as index.html where the content was reduced. This is due to a more efficient react approach in the index.js file were the prior html code was transformed into jsx.